+++ frontend/jest.config.cjs
@@ -2,10 +2,10 @@ module.exports = {
   testEnvironment: 'jsdom',
   setupFilesAfterEnv: ['<rootDir>/jest.setup.cjs'],
   moduleNameMapper: {
-    '\.(css|less|scss|sass)$': 'identity-obj-proxy'
+    '\.(css|less|scss|sass)$': 'identity-obj-proxy',
   },
   transform: {
-    '^.+\.[tj]sx?$': 'babel-jest'
+    '^.+\.[tj]sx?$': 'babel-jest',
   },
-  moduleFileExtensions: ['js', 'jsx', 'json']
+  moduleFileExtensions: ['ts', 'tsx', 'json'],
 };
