+++ frontend/package.json
@@ -18,9 +18,11 @@
   "devDependencies": {
     "@babel/preset-env": "^7.25.4",
     "@babel/preset-react": "^7.24.7",
-    "@testing-library/jest-dom": "^6.4.6",
+    "@babel/preset-typescript": "^7.24.7",
+    "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^14.3.1",
     "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.12",
     "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.2",
     "@vitejs/plugin-react": "^4.3.1",
@@ -28,6 +30,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "typescript": "^5.6.3",
     "vite": "^5.4.2"
   }
 }
