+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
-import React from "react";
+import { FC } from 'react';
 
-function App() {
+const App: FC = () => {
   return <div className="App">Hello World</div>;
-}
+};
 
 export default App;
