--- frontend/src/api.js
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const api = axios.create({
-  baseURL: import.meta.env.VITE_API_URL ?? "http://localhost:4000",
-});
-
-export default api;
