+++ frontend/src/api.ts
@@ -1,0 +1,9 @@
+import axios from 'axios';
+
+const api = axios.create({
+  baseURL: import.meta.env.VITE_API_URL ?? 'http://localhost:4000',
+});
+
+export type ApiClient = typeof api;
+
+export default api;
