--- frontend/vite.config.js
@@ -1,12 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-
-export default defineConfig({
-  plugins: [react()],
-  server: {
-    port: 3000
-  },
-  preview: {
-    port: 3000
-  }
-});
